Restructure as parametrization
