+++ nevergrad/parametrization/parameter.py
@@ -5,8 +5,8 @@
 
 # pylint: disable=unused-import
 # import with "as" to explicitely allow reexport (mypy)
-from .core3 import Parameter as Parameter
-from .core3 import Dict as Dict
+from .core import Parameter as Parameter
+from .core import Dict as Dict
 from .container import Tuple as Tuple
 from .container import Instrumentation as Instrumentation
 from .data import Array as Array
